pyproject.toml
[project]
name = "vibe-code-cli-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "colorama>=0.4.6",
    "google-generativeai>=0.8.5",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
]
